Backend improvements
You can reset the amount (volume/mass) of a container

Transactions now have types, which shows whether it's a standard transaction, or an amount reset, etc.

I didn't work on styling sorry

Also changed the Remove button when creating a transaction -- it no longer negates the slider value but rather changes a variable which is not displayed. I.e. it still works but doesn't convolute UI